refactor(inspection-service): extract inspection-analytics sub-service
Move getCounts / getDefectStats / getDefectStatsBatch / getDashboardBuckets /
getRepairList / getObserveProgress into InspectionAnalyticsService. Methods
that need the full resolved report call back through the facade (getReportData)
via an injected facade reference — cross-service dependency made explicit. The
facade delegates each method; moved code is byte-identical.
